+++ public/all.php
@@ -1,2 +1,3 @@
 <?php
+require_once 'hoo-functions.php';
 require_once 'hoo-checkout.php';