+++ woocommerce-edje.php
@@ -5,7 +5,7 @@ Description: Library that helps customize WooCommerce. Designed to work with Tim
 Plugin URI: http://github.com/hrsetyono/woocommerce-edje
 Author: The Syne Studio
 Author URI: http://thesyne.com/
-Version: 1.0.0-beta
+Version: 1.0.1-beta
 */
 
 // exit if accessed directly
@@ -25,22 +25,20 @@ run_hoo();
 // Main portal for calling all methods
 class Hoo {
   function __construct() {
-    global $post;
-
     // disable everything if WC not installed
     include_once(ABSPATH . 'wp-admin/includes/plugin.php');
-    if(!is_plugin_active( 'woocommerce/woocommerce.php' ) ) {
-      var_dump('wc not active');
+    if(!is_plugin_active('woocommerce/woocommerce.php') ) {
       return false;
     }
 
     // for product edit only
+    global $post;
     if(isset($post->post_type) && $post->post_type == 'product') {
       add_action('admin_init', array($this, 'init_admin') );
     }
 
-    // for front-end woocommerce page
-    // $this->init_public();
+    // front end woocommerce page
+    add_action('template_redirect', array($this, 'init_public'));
   }
 
   function init_admin() {
@@ -50,6 +48,8 @@ class Hoo {
   }
 
   function init_public() {
-    $hoo_checkout = new Hoo_Checkout();
+    if(is_checkout() ) {
+      $hoo_checkout = new Hoo_Checkout();
+    }
   }
 }